fixed up optimization waypoint object
